Project structure rework
