+++ package.json
@@ -2,22 +2,22 @@
   "name": "markdown-proofreader",
   "version": "0.0.4",
   "description": "",
-  "main": "src/index.js",
+  "main": "src/cli.js",
   "bin": {
-    "markdown-proofreader": "./src/index.js"
+    "markdown-proofreader": "./src/cli.js"
   },
   "scripts": {
     "test": "npm run lint && npm run test:unit",
     "test:unit": "jest __test__/unit/**",
     "test:e2e": "jest __test__/e2e/**",
-    "start:dev": "node src/index.js",
+    "start:dev": "node src/cli.js",
     "lint": "eslint --ext .js src/ __test__/"
   },
   "files": [
     "src"
   ],
   "keywords": [],
-  "author": "",
+  "author": "Volubyl",
   "license": "MIT",
   "devDependencies": {
     "eslint": "^6.8.0",
